refactor(direct-control): extract battlefield scan source
Move the battlefield scan embedded App UI source into src/play/tactical/battlefield.ts while keeping the public wrapper and command builder in the facade. Preserve the conservative relationship-proof policy and keep tactical helper logic local to the battlefield owner.

Record the tactical read atom status in the support modularization OpenSpec workstream. This is source ownership proof only and does not claim live-game runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- tactical-reads.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
